Loaded Pandas DFs to DB, changed folder names, moved functions to utils.py file
